View the codebook of a dataset or selected variables.

codebook() is a function to view the codebook of a dataset or selected variables.

Usage

codebook(.data, ...)

Arguments

.data

The input data (data frame or tibble).

...

<tidy-select> One or more unquoted expressions separated by commas. Variable names can be used as if they were positions in the data frame, so expressions like x:y can be used to select a range of variables. This argument can be omitted.

Value

An object of the same type as .data.

Examples

starwars
#> # A tibble: 87 × 14
#>    name     height  mass hair_color skin_color eye_color birth_year sex   gender
#>    <chr>     <int> <dbl> <chr>      <chr>      <chr>          <dbl> <chr> <chr> 
#>  1 Luke Sk…    172    77 blond      fair       blue            19   male  mascu…
#>  2 C-3PO       167    75 NA         gold       yellow         112   none  mascu…
#>  3 R2-D2        96    32 NA         white, bl… red             33   none  mascu…
#>  4 Darth V…    202   136 none       white      yellow          41.9 male  mascu…
#>  5 Leia Or…    150    49 brown      light      brown           19   fema… femin…
#>  6 Owen La…    178   120 brown, gr… light      blue            52   male  mascu…
#>  7 Beru Wh…    165    75 brown      light      blue            47   fema… femin…
#>  8 R5-D4        97    32 NA         white, red red             NA   none  mascu…
#>  9 Biggs D…    183    84 black      light      brown           24   male  mascu…
#> 10 Obi-Wan…    182    77 auburn, w… fair       blue-gray       57   male  mascu…
#> # ℹ 77 more rows
#> # ℹ 5 more variables: homeworld <chr>, species <chr>, films <list>,
#> #   vehicles <list>, starships <list>
codebook(starwars)
#> # A tibble: 14 × 4
#>    variable   type          n unique
#>    <chr>      <chr>     <int>  <int>
#>  1 name       character    87     87
#>  2 height     integer      81     45
#>  3 mass       double       59     38
#>  4 hair_color character    82     12
#>  5 skin_color character    87     31
#>  6 eye_color  character    87     15
#>  7 birth_year double       43     36
#>  8 sex        character    83      4
#>  9 gender     character    83      2
#> 10 homeworld  character    77     48
#> 11 species    character    83     37
#> 12 films      list         87     24
#> 13 vehicles   list         87     11
#> 14 starships  list         87     17
codebook(starwars, 1:4)
#> # A tibble: 4 × 4
#>   variable   type          n unique
#>   <chr>      <chr>     <int>  <int>
#> 1 name       character    87     87
#> 2 height     integer      81     45
#> 3 mass       double       59     38
#> 4 hair_color character    82     12
codebook(starwars, ends_with("color"))
#> # A tibble: 3 × 4
#>   variable   type          n unique
#>   <chr>      <chr>     <int>  <int>
#> 1 hair_color character    82     12
#> 2 skin_color character    87     31
#> 3 eye_color  character    87     15
codebook(starwars, where(is.numeric))
#> # A tibble: 3 × 4
#>   variable   type        n unique
#>   <chr>      <chr>   <int>  <int>
#> 1 height     integer    81     45
#> 2 mass       double     59     38
#> 3 birth_year double     43     36
lifeexp
#> # A tibble: 68 × 6
#>    region               country                popgrowth  lexp gnppc safewater
#>    <dbl+lbl>            <chr>                      <dbl> <dbl> <dbl>     <dbl>
#>  1 1 [Europe & C. Asia] Albania                    1.20     72   810        76
#>  2 1 [Europe & C. Asia] Armenia                    1.10     74   460        NA
#>  3 1 [Europe & C. Asia] Austria                    0.400    79 26830        NA
#>  4 1 [Europe & C. Asia] Azerbaijan                 1.40     71   480        NA
#>  5 1 [Europe & C. Asia] Belarus                    0.300    68  2180        NA
#>  6 1 [Europe & C. Asia] Belgium                    0.200    78 25380        NA
#>  7 1 [Europe & C. Asia] Bosnia and Herzegovina    -0.5      73    NA        NA
#>  8 1 [Europe & C. Asia] Bulgaria                  -0.400    71  1220        NA
#>  9 1 [Europe & C. Asia] Croatia                   -0.100    73  4620        63
#> 10 1 [Europe & C. Asia] Czech Republic             0        75  5150        NA
#> # ℹ 58 more rows
codebook(lifeexp)
#> # A tibble: 6 × 5
#>   variable  type             n unique label                   
#>   <chr>     <chr>        <int>  <int> <chr>                   
#> 1 region    double+label    68      3 Region                  
#> 2 country   character       68     68 Country                 
#> 3 popgrowth double          68     30 Avg. annual % growth    
#> 4 lexp      double          68     18 Life expectancy at birth
#> 5 gnppc     double          63     62 GNP per capita          
#> 6 safewater double          40     29 Safe water