Tabulate columns

Enhanced count().

tab() - Tabulates variable(s) simultaneously, similar to count().

tab1() - Tabulates variable(s) one by one.

tab2() - Cross-tabulates two variables.

fre() - Adds “Total” row(s) to the output of tab().

fre1() - Adds “Total” row(s) to the output of tab1().

fre2() - Adds a “Total” row and a “total” column to the output of tab2().

Overall, tab() functions are suitable for subsequent data analysis, while fre() functions are suitable for printing.

If the output tibble has many rows, consider using print() or s_print() to display the output.

Usage

tab(.data, ..., .desc = FALSE)

tab1(.data, ..., .desc = FALSE, .append = FALSE)

tab2(.data, ..., .flip = FALSE)

fre(.data, ..., .desc = FALSE)

fre1(.data, ..., .desc = FALSE, .append = FALSE)

fre2(.data, ..., .flip = FALSE)

Arguments

.data

A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).

...

<tidy-select> or <data-masking> Variables to tabulate.

Both tidyselect (e.g., starts_with()) and data masking (e.g., x_sq = x^2) are supported. See examples below.

.desc

If TRUE, will show the largest groups at the top.

.append

If TRUE, will return a single tibble instead of a list of tibbles.

.flip

If TRUE, will transpose the output tibble.

Value

An object of the same type as .data.

Examples

tab(starwars, 8)
#> # A tibble: 5 × 6
#>   sex                n percent   cum valid valid_cum
#>   <chr>          <int>   <dbl> <dbl> <dbl>     <dbl>
#> 1 female            16   18.4   18.4 19.3       19.3
#> 2 hermaphroditic     1    1.15  19.5  1.20      20.5
#> 3 male              60   69.0   88.5 72.3       92.8
#> 4 none               6    6.90  95.4  7.23     100  
#> 5 NA                 4    4.60 100   NA         NA  
tab(starwars, sex)
#> # A tibble: 5 × 6
#>   sex                n percent   cum valid valid_cum
#>   <chr>          <int>   <dbl> <dbl> <dbl>     <dbl>
#> 1 female            16   18.4   18.4 19.3       19.3
#> 2 hermaphroditic     1    1.15  19.5  1.20      20.5
#> 3 male              60   69.0   88.5 72.3       92.8
#> 4 none               6    6.90  95.4  7.23     100  
#> 5 NA                 4    4.60 100   NA         NA  
tab(starwars, sex, .desc = TRUE)
#> # A tibble: 5 × 6
#>   sex                n percent   cum valid valid_cum
#>   <chr>          <int>   <dbl> <dbl> <dbl>     <dbl>
#> 1 male              60   69.0   69.0 72.3       72.3
#> 2 female            16   18.4   87.4 19.3       91.6
#> 3 none               6    6.90  94.3  7.23      98.8
#> 4 hermaphroditic     1    1.15  95.4  1.20     100  
#> 5 NA                 4    4.60 100   NA         NA  
fre(starwars, sex)
#> # A tibble: 7 × 6
#>   sex                n percent   cum  valid valid_cum
#>   <chr>          <int>   <dbl> <dbl>  <dbl>     <dbl>
#> 1 female            16   18.4   18.4  19.3       19.3
#> 2 hermaphroditic     1    1.15  19.5   1.20      20.5
#> 3 male              60   69.0   88.5  72.3       92.8
#> 4 none               6    6.90  95.4   7.23     100  
#> 5 Valid Total       83   95.4   NA   100         NA  
#> 6 NA                 4    4.60 100    NA         NA  
#> 7 Total             87  100     NA    NA         NA  
fre(starwars, sex, .desc = TRUE)
#> # A tibble: 7 × 6
#>   sex                n percent   cum  valid valid_cum
#>   <chr>          <int>   <dbl> <dbl>  <dbl>     <dbl>
#> 1 male              60   69.0   69.0  72.3       72.3
#> 2 female            16   18.4   87.4  19.3       91.6
#> 3 none               6    6.90  94.3   7.23      98.8
#> 4 hermaphroditic     1    1.15  95.4   1.20     100  
#> 5 Valid Total       83   95.4   NA   100         NA  
#> 6 NA                 4    4.60 100    NA         NA  
#> 7 Total             87  100     NA    NA         NA  

tab(starwars, sex, gender)
#> # A tibble: 6 × 7
#>   sex            gender        n percent   cum valid valid_cum
#>   <chr>          <chr>     <int>   <dbl> <dbl> <dbl>     <dbl>
#> 1 female         feminine     16   18.4   18.4 19.3       19.3
#> 2 hermaphroditic masculine     1    1.15  19.5  1.20      20.5
#> 3 male           masculine    60   69.0   88.5 72.3       92.8
#> 4 none           feminine      1    1.15  89.7  1.20      94.0
#> 5 none           masculine     5    5.75  95.4  6.02     100  
#> 6 NA             NA            4    4.60 100   NA         NA  
tab(starwars, sex == "male")
#> # A tibble: 3 × 6
#>   `sex == "male"`     n percent   cum valid valid_cum
#>   <lgl>           <int>   <dbl> <dbl> <dbl>     <dbl>
#> 1 FALSE              23   26.4   26.4  27.7      27.7
#> 2 TRUE               60   69.0   95.4  72.3     100  
#> 3 NA                  4    4.60 100    NA        NA  
tab(starwars, male = sex == "male")
#> # A tibble: 3 × 6
#>   male      n percent   cum valid valid_cum
#>   <lgl> <int>   <dbl> <dbl> <dbl>     <dbl>
#> 1 FALSE    23   26.4   26.4  27.7      27.7
#> 2 TRUE     60   69.0   95.4  72.3     100  
#> 3 NA        4    4.60 100    NA        NA  
tab(starwars, ends_with("color"))
#> # A tibble: 67 × 8
#>    hair_color    skin_color eye_color     n percent   cum valid valid_cum
#>    <chr>         <chr>      <chr>     <int>   <dbl> <dbl> <dbl>     <dbl>
#>  1 auburn        fair       blue          1    1.15  1.15  1.22      1.22
#>  2 auburn, grey  fair       blue          1    1.15  2.30  1.22      2.44
#>  3 auburn, white fair       blue-gray     1    1.15  3.45  1.22      3.66
#>  4 black         blue, grey yellow        1    1.15  4.60  1.22      4.88
#>  5 black         brown      brown         1    1.15  5.75  1.22      6.10
#>  6 black         dark       brown         3    3.45  9.20  3.66      9.76
#>  7 black         dark       dark          1    1.15 10.3   1.22     11.0 
#>  8 black         fair       brown         2    2.30 12.6   2.44     13.4 
#>  9 black         light      brown         1    1.15 13.8   1.22     14.6 
#> 10 black         tan        brown         2    2.30 16.1   2.44     17.1 
#> # ℹ 57 more rows
tab(starwars, s_match("s*e"))
#> # A tibble: 1 × 5
#>       n percent   cum valid valid_cum
#>   <int>   <dbl> <dbl> <dbl>     <dbl>
#> 1    87     100   100   100       100
tab(starwars)
#> # A tibble: 87 × 19
#>    name     height  mass hair_color skin_color eye_color birth_year sex   gender
#>    <chr>     <int> <dbl> <chr>      <chr>      <chr>          <dbl> <chr> <chr> 
#>  1 Ackbar      180  83   none       brown mot… orange          41   male  mascu…
#>  2 Anakin …    188  84   blond      fair       blue            41.9 male  mascu…
#>  3 Ayla Se…    178  55   none       blue       hazel           48   fema… femin…
#>  4 Barriss…    166  50   black      yellow     blue            40   fema… femin…
#>  5 Beru Wh…    165  75   brown      light      blue            47   fema… femin…
#>  6 Biggs D…    183  84   black      light      brown           24   male  mascu…
#>  7 Boba Fe…    183  78.2 black      fair       brown           31.5 male  mascu…
#>  8 Bossk       190 113   none       green      red             53   male  mascu…
#>  9 Chewbac…    228 112   brown      unknown    blue           200   male  mascu…
#> 10 Darth M…    175  80   none       red        yellow          54   male  mascu…
#> # ℹ 77 more rows
#> # ℹ 10 more variables: homeworld <chr>, species <chr>, films <list>,
#> #   vehicles <list>, starships <list>, n <int>, percent <dbl>, cum <dbl>,
#> #   valid <dbl>, valid_cum <dbl>

tab1(starwars, 1:3)
#> $name
#> # A tibble: 87 × 6
#>    value                   n percent   cum valid valid_cum
#>    <chr>               <int>   <dbl> <dbl> <dbl>     <dbl>
#>  1 Ackbar                  1    1.15  1.15  1.15      1.15
#>  2 Adi Gallia              1    1.15  2.30  1.15      2.30
#>  3 Anakin Skywalker        1    1.15  3.45  1.15      3.45
#>  4 Arvel Crynyd            1    1.15  4.60  1.15      4.60
#>  5 Ayla Secura             1    1.15  5.75  1.15      5.75
#>  6 BB8                     1    1.15  6.90  1.15      6.90
#>  7 Bail Prestor Organa     1    1.15  8.05  1.15      8.05
#>  8 Barriss Offee           1    1.15  9.20  1.15      9.20
#>  9 Ben Quadinaros          1    1.15 10.3   1.15     10.3 
#> 10 Beru Whitesun lars      1    1.15 11.5   1.15     11.5 
#> # ℹ 77 more rows
#> 
#> $height
#> # A tibble: 46 × 6
#>    value     n percent   cum valid valid_cum
#>    <int> <int>   <dbl> <dbl> <dbl>     <dbl>
#>  1    66     1    1.15  1.15  1.23      1.23
#>  2    79     1    1.15  2.30  1.23      2.47
#>  3    88     1    1.15  3.45  1.23      3.70
#>  4    94     1    1.15  4.60  1.23      4.94
#>  5    96     2    2.30  6.90  2.47      7.41
#>  6    97     1    1.15  8.05  1.23      8.64
#>  7   112     1    1.15  9.20  1.23      9.88
#>  8   122     1    1.15 10.3   1.23     11.1 
#>  9   137     1    1.15 11.5   1.23     12.3 
#> 10   150     2    2.30 13.8   2.47     14.8 
#> # ℹ 36 more rows
#> 
#> $mass
#> # A tibble: 39 × 6
#>    value     n percent   cum valid valid_cum
#>    <dbl> <int>   <dbl> <dbl> <dbl>     <dbl>
#>  1    15     1    1.15  1.15  1.69      1.69
#>  2    17     1    1.15  2.30  1.69      3.39
#>  3    20     1    1.15  3.45  1.69      5.08
#>  4    32     2    2.30  5.75  3.39      8.47
#>  5    40     1    1.15  6.90  1.69     10.2 
#>  6    45     2    2.30  9.20  3.39     13.6 
#>  7    48     2    2.30 11.5   3.39     16.9 
#>  8    49     1    1.15 12.6   1.69     18.6 
#>  9    50     2    2.30 14.9   3.39     22.0 
#> 10    55     2    2.30 17.2   3.39     25.4 
#> # ℹ 29 more rows
#> 
tab1(starwars, ends_with("color"))
#> $hair_color
#> # A tibble: 13 × 6
#>    value             n percent    cum valid valid_cum
#>    <chr>         <int>   <dbl>  <dbl> <dbl>     <dbl>
#>  1 auburn            1    1.15   1.15  1.22      1.22
#>  2 auburn, grey      1    1.15   2.30  1.22      2.44
#>  3 auburn, white     1    1.15   3.45  1.22      3.66
#>  4 black            13   14.9   18.4  15.9      19.5 
#>  5 blond             3    3.45  21.8   3.66     23.2 
#>  6 blonde            1    1.15  23.0   1.22     24.4 
#>  7 brown            18   20.7   43.7  22.0      46.3 
#>  8 brown, grey       1    1.15  44.8   1.22     47.6 
#>  9 grey              1    1.15  46.0   1.22     48.8 
#> 10 none             37   42.5   88.5  45.1      93.9 
#> 11 unknown           1    1.15  89.7   1.22     95.1 
#> 12 white             4    4.60  94.3   4.88    100   
#> 13 NA                5    5.75 100    NA        NA   
#> 
#> $skin_color
#> # A tibble: 31 × 6
#>    value                   n percent   cum valid valid_cum
#>    <chr>               <int>   <dbl> <dbl> <dbl>     <dbl>
#>  1 blue                    2    2.30  2.30  2.30      2.30
#>  2 blue, grey              2    2.30  4.60  2.30      4.60
#>  3 brown                   4    4.60  9.20  4.60      9.20
#>  4 brown mottle            1    1.15 10.3   1.15     10.3 
#>  5 brown, white            1    1.15 11.5   1.15     11.5 
#>  6 dark                    6    6.90 18.4   6.90     18.4 
#>  7 fair                   17   19.5  37.9  19.5      37.9 
#>  8 fair, green, yellow     1    1.15 39.1   1.15     39.1 
#>  9 gold                    1    1.15 40.2   1.15     40.2 
#> 10 green                   6    6.90 47.1   6.90     47.1 
#> # ℹ 21 more rows
#> 
#> $eye_color
#> # A tibble: 15 × 6
#>    value             n percent   cum valid valid_cum
#>    <chr>         <int>   <dbl> <dbl> <dbl>     <dbl>
#>  1 black            10   11.5   11.5 11.5       11.5
#>  2 blue             19   21.8   33.3 21.8       33.3
#>  3 blue-gray         1    1.15  34.5  1.15      34.5
#>  4 brown            21   24.1   58.6 24.1       58.6
#>  5 dark              1    1.15  59.8  1.15      59.8
#>  6 gold              1    1.15  60.9  1.15      60.9
#>  7 green, yellow     1    1.15  62.1  1.15      62.1
#>  8 hazel             3    3.45  65.5  3.45      65.5
#>  9 orange            8    9.20  74.7  9.20      74.7
#> 10 pink              1    1.15  75.9  1.15      75.9
#> 11 red               5    5.75  81.6  5.75      81.6
#> 12 red, blue         1    1.15  82.8  1.15      82.8
#> 13 unknown           3    3.45  86.2  3.45      86.2
#> 14 white             1    1.15  87.4  1.15      87.4
#> 15 yellow           11   12.6  100   12.6      100  
#> 
tab1(starwars, ends_with("color"), .append = TRUE)
#> # A tibble: 59 × 7
#>    variable   value             n percent   cum valid valid_cum
#>    <chr>      <chr>         <int>   <dbl> <dbl> <dbl>     <dbl>
#>  1 hair_color auburn            1    1.15  1.15  1.22      1.22
#>  2 hair_color auburn, grey      1    1.15  2.30  1.22      2.44
#>  3 hair_color auburn, white     1    1.15  3.45  1.22      3.66
#>  4 hair_color black            13   14.9  18.4  15.9      19.5 
#>  5 hair_color blond             3    3.45 21.8   3.66     23.2 
#>  6 hair_color blonde            1    1.15 23.0   1.22     24.4 
#>  7 hair_color brown            18   20.7  43.7  22.0      46.3 
#>  8 hair_color brown, grey       1    1.15 44.8   1.22     47.6 
#>  9 hair_color grey              1    1.15 46.0   1.22     48.8 
#> 10 hair_color none             37   42.5  88.5  45.1      93.9 
#> # ℹ 49 more rows
fre1(starwars, ends_with("color"), .append = TRUE)
#> # A tibble: 63 × 7
#>    variable   value             n percent   cum valid valid_cum
#>    <chr>      <chr>         <int>   <dbl> <dbl> <dbl>     <dbl>
#>  1 hair_color auburn            1    1.15  1.15  1.22      1.22
#>  2 hair_color auburn, grey      1    1.15  2.30  1.22      2.44
#>  3 hair_color auburn, white     1    1.15  3.45  1.22      3.66
#>  4 hair_color black            13   14.9  18.4  15.9      19.5 
#>  5 hair_color blond             3    3.45 21.8   3.66     23.2 
#>  6 hair_color blonde            1    1.15 23.0   1.22     24.4 
#>  7 hair_color brown            18   20.7  43.7  22.0      46.3 
#>  8 hair_color brown, grey       1    1.15 44.8   1.22     47.6 
#>  9 hair_color grey              1    1.15 46.0   1.22     48.8 
#> 10 hair_color none             37   42.5  88.5  45.1      93.9 
#> # ℹ 53 more rows
fre1(starwars, ends_with("color"), .append = TRUE) %>% s_print()
#> # A tibble: 63 × 7
#>    variable   value             n percent    cum valid valid_cum
#>    <chr>      <chr>         <int>   <dbl>  <dbl> <dbl>     <dbl>
#>  1 hair_color auburn            1    1.15   1.15  1.22      1.22
#>  2 hair_color auburn, grey      1    1.15   2.30  1.22      2.44
#>  3 hair_color auburn, white     1    1.15   3.45  1.22      3.66
#>  4 hair_color black            13   14.9   18.4  15.9      19.5 
#>  5 hair_color blond             3    3.45  21.8   3.66     23.2 
#> 59 eye_color  red, blue         1    1.15  82.8   1.15     82.8 
#> 60 eye_color  unknown           3    3.45  86.2   3.45     86.2 
#> 61 eye_color  white             1    1.15  87.4   1.15     87.4 
#> 62 eye_color  yellow           11   12.6  100    12.6     100   
#> 63 eye_color  Total            87  100     NA    NA        NA   
#> # ℹ 53 more rows in the middle
#> # ℹ Use `s_print(n = ...)` to see more rows

tab2(starwars, sex, gender)
#> # A tibble: 5 × 4
#>   `sex \\ gender` feminine masculine  `NA`
#>   <chr>              <int>     <int> <int>
#> 1 female                16         0     0
#> 2 hermaphroditic         0         1     0
#> 3 male                   0        60     0
#> 4 none                   1         5     0
#> 5 NA                     0         0     4
tab2(starwars, height, gender)
#> # A tibble: 46 × 4
#>    `height \\ gender` masculine feminine  `NA`
#>                 <int>     <int>    <int> <int>
#>  1                 66         1        0     0
#>  2                 79         1        0     0
#>  3                 88         1        0     0
#>  4                 94         1        0     0
#>  5                 96         1        1     0
#>  6                 97         1        0     0
#>  7                112         1        0     0
#>  8                122         1        0     0
#>  9                137         1        0     0
#> 10                150         0        2     0
#> # ℹ 36 more rows
fre2(starwars, height, gender)
#> # A tibble: 47 × 5
#>    `height \\ gender` masculine feminine  `NA` total
#>    <chr>                  <int>    <int> <int> <int>
#>  1 66                         1        0     0     1
#>  2 79                         1        0     0     1
#>  3 88                         1        0     0     1
#>  4 94                         1        0     0     1
#>  5 96                         1        1     0     2
#>  6 97                         1        0     0     1
#>  7 112                        1        0     0     1
#>  8 122                        1        0     0     1
#>  9 137                        1        0     0     1
#> 10 150                        0        2     0     2
#> # ℹ 37 more rows
fre2(starwars, height, gender) %>% s_print()
#> # A tibble: 47 × 5
#>    `height \\ gender` masculine feminine  `NA` total
#>    <chr>                  <int>    <int> <int> <int>
#>  1 66                         1        0     0     1
#>  2 79                         1        0     0     1
#>  3 88                         1        0     0     1
#>  4 94                         1        0     0     1
#>  5 96                         1        1     0     2
#> 43 229                        1        0     0     1
#> 44 234                        1        0     0     1
#> 45 264                        1        0     0     1
#> 46 NA                         4        1     1     6
#> 47 total                     66       17     4    87
#> # ℹ 37 more rows in the middle
#> # ℹ Use `s_print(n = ...)` to see more rows